🔧 chore: Move share setting components
