+++ src/app/(main)/settings/_layout/Desktop/index.tsx
@@ -6,11 +6,11 @@ import { memo, useRef } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Flexbox } from 'react-layout-kit';
 
+import SettingContainer from '@/features/Setting//SettingContainer';
+import Footer from '@/features/Setting/Footer';
 import { useActiveSettingsKey } from '@/hooks/useActiveSettingsKey';
 import { SettingsTabs } from '@/store/global/initialState';
 
-import Footer from '../../features/Footer';
-import SettingContainer from '../../features/SettingContainer';
 import { LayoutProps } from '../type';
 import Header from './Header';
 import SideBar from './SideBar';
